requirements.txt
django
openai
python-dotenv
requests